+++ nengo/tests/options.py
@@ -24,3 +24,5 @@ def pytest_addoption(parser):
         help='Also run slow tests.')
     parser.addoption('--seed-offset', nargs=1, type=int, default=0,
                      help="Specify offset of the seed values used in tests.")
+    parser.addoption('--spa', action='store_true', default=False,
+                     help='Run deprecated SPA tests')